requirements.txt
absl-py==2.1.0
accelerate==0.24.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
bitsandbytes==0.35.4
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
decord==0.6.0
diffusers==0.19.0
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.0
grpcio==1.65.0
huggingface-hub==0.17.3
humanfriendly==10.0
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib-metadata==8.0.0
ipython==8.26.0
jedi==0.19.1
kiwisolver==1.4.5
markdown==3.6
matplotlib==3.9.1
matplotlib-inline==0.1.7
modelcards==0.1.6
omegaconf==2.3.0
onnxruntime==1.18.1
onnxruntime-gpu==1.18.1
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyav==12.2.0
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
regex==2024.5.15
safetensors==0.4.3
six==1.16.0
stack-data==0.6.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tokenizers==0.14.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.35.0
triton==2.3.1
wcwidth==0.2.13
werkzeug==3.0.3
zipp==3.19.2
scipy==1.14.0
scikit-learn==1.2.2
nltk==3.8.1
timm==0.6.7
scikit-image==0.24.0
gdown==5.1.0
basicsr-fixed